+++ lib/docking.dart
@@ -10,5 +10,7 @@ export 'src/on_item_close.dart';
 export 'src/docking_buttons_builder.dart';
 export 'src/docking_icons.dart';
 export 'src/docking.dart';
+export 'src/theme/docking_theme.dart';
+export 'src/theme/docking_theme_data.dart';
 export 'package:multi_split_view/multi_split_view.dart';
 export 'package:tabbed_view/tabbed_view.dart';